+++ converters/pdf/src/lib.rs
@@ -191,6 +191,7 @@ impl Processor<'_> {
         let mut visitor = PdfVisitor::new(
             processor,
             assets,
+            theme.heading,
             !doc.toc_entries.is_empty(),
             diagnostics.reborrow(),
         );